finished permissions

+++ website/main/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth import login, logout, authenticate
 from django.contrib.auth.decorators import login_required, permission_required
+from django.contrib.auth.models import User, Group
 
 from .forms import RegistrationForm, PostForm
 from .models import Post
@@ -12,10 +13,19 @@ def home(request):
     posts = Post.objects.all()
     if request.method == 'POST':
         post_id = request.POST.get('post-id')
-        post = Post.objects.filter(id=int(post_id)).first()
-        if post and (post.author == request.user or request.user.has_perm("main.delete_post")):
-            post.delete()
-        
+        if post_id:
+            post = Post.objects.filter(id=int(post_id)).first()
+            if post and (post.author == request.user or request.user.has_perm("main.delete_post")):
+                post.delete()
+        user_id = request.POST.get('user-id')
+        if user_id:
+            user = User.objects.filter(id=user_id).first()
+            if user and request.user.is_staff:
+                group = Group.objects.get(name='default')
+                group.user_set.remove(user)
+
+                group = Group.objects.get(name='mod')
+                group.user_set.remove(user)
     return render(request, "main/home.html", {'posts': posts})
 
 def sign_up(request):
